Grapple attack pt.1
Grapple attack script and animations has been created and imported. Still this feature is not finished.
