+++ True_REKT/Assets/Animations/No_Gun/Player_Grapple_01.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a25331761d98f84fbca6d02abe35645
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
