+++ True_REKT/Assets/Animations/Zombie/Zombie_Attack_Lunge_01.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 631c8ee45d0a7014585c428a735e7908
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
