+++ True_REKT/Assets/Animations/Zombie/Zombie_Grapple.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 00208870cd8221841bd4166463111d4e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
