--- True_REKT/Assets/Data/Zombie Data/Zombie_Arm_attack.asset
@@ -1,20 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2fb9b12ef3ef17a42bdea603100fa1d7, type: 3}
-  m_Name: Zombie_Arm_attack
-  m_EditorClassIdentifier: 
-  attackAnimation: Zombie_Attack
-  attackCooldown: 5
-  maximumAttackAngle: 20
-  minimumAttackAngle: -20
-  maximumAttackDistance: 0.5
-  minimumAttackDistance: 2
