+++ True_REKT/Assets/Data/Zombie Data/Zombie_Lunge_Attack.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7dfe030ab6d5aa49981df730ce35d38
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
