--- True_REKT/Assets/__Scripts/Player/AnimationManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: be03253067f26c7408b56897ae94b705
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
