+++ True_REKT/Assets/__Scripts/Player/PlayerAnimationManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e178d2a88139f85418e201e5786998fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
