+++ True_REKT/Assets/__Scripts/Zombie/ZombieGrappleCollider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30108df0c3539784ea6b48733b4bd145
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
